Finised move to proxy/utils
